groupserv: Fix incorrect behaviour for flags +*
In the GroupServ flags parser, asterisk is handled by setting the flags
to GA_ALL. This does not preserve flags that are not part of GA_ALL,
such as the founder flag. As a result, a sole founder of a group cannot
set flags +* on themselves.

This patch amends gs_flags_parser to preserve existing flags on +*
except for GA_BAN.