groupserv: Fix inconsistencies with FLAGS
At present, a GroupServ FLAGS command issued on an account with flags in
the group in question, without a change in flags, does not invoke
command_fail with the fault_nochange fault reason.

Additionally, if a FLAGS command is issued on an account that does not
exist in the group in question, ultimately without any flags being
applied, the account is still added to the group access list. If a FLAGS
command meeting these criteria is issued a second time on the same
account, the account is removed from the group access list as is
expected for an account no longer holding flags in the group.

This patch fixes both issues described above, the latter of which was
first addressed by @Renegade334 in #432.